services/rl-policy-service/requirements.txt
fastapi
uvicorn
kafka-python
numpy
torch
gymnasium
redis
pydantic
